move App and Pet tests in __tests__ folder
